+++ akashic-storage/src/createValkeyConnection.ts
@@ -16,8 +16,6 @@ export const createValkeyConnection = async (
         inflightRequestsLimit: parseInt(
             process.env.INFLIGHT_REQUEST_LIMIT ?? "1000",
         ),
-        requestTimeout: process.env.VALKEY_REQUEST_TIMEOUT_MS
-            ? parseInt(process.env.VALKEY_REQUEST_TIMEOUT_MS)
-            : undefined,
+        requestTimeout: parseInt(process.env.REQUEST_TIMEOUT ?? "250"),
         readFrom: "preferReplica",
     });